requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
load-dotenv==0.1.0
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
telebot==0.0.5
urllib3==2.3.0
